requirements.txt
viam-sdk
smbus-cffi